refactor(vetting): Phase 3b-2 — ApplicationUser.VettingStatus entity int → enum
Final backend sub-commit of Phase 3. Converts ApplicationUser.VettingStatus
from int to the VettingStatus enum, cascading through every consumer of
the entity, and removes every temporary cast / integer constant /
magic-number comparison that Phase 2 and Phase 3b-1 had introduced as
migration scaffolding.

## Entity

- `ApplicationUser.cs` — VettingStatus property type changed from int
  to VettingStatus enum. Default initializer is now
  `VettingStatus.UnderReview` (was `0`). The `IsVetted` computed property
  now compares against `VettingStatus.Approved` instead of magic `== 3`.

## EF Core migration

- NEW: `Migrations/20260411041245_Phase3b2VettingStatusEnum.cs`
  EF Core detected the CLR type change but generated an empty migration
  because enums are stored as their underlying int type by default —
  the database column remains `integer`, no data migration needed.

  Added a `CHECK (VettingStatus BETWEEN 0 AND 6)` constraint in the
  migration's Up() (reversible in Down()) per the Phase 3b-1 code review
  S2 finding. This defends against the C# quirk where `(VettingStatus)99`
  is legal and would otherwise be silently written to the DB.

  Constraint name: `CK_Users_VettingStatus_Range` (the Users table, not
  AspNetUsers — Identity is customized via `ToTable("Users")`).

## LINQ-to-SQL queries (4 files, ~13 sites)

All magic-number comparisons and `(int)VettingStatus.X` casts in LINQ
expressions that translate to SQL replaced with direct enum comparisons
(EF Core handles enum → int translation natively for the underlying
column type):

- `EmailTemplateService.cs` — 6 segment-filter queries plus
  GetVettingStatusDisplay helper signature int → enum
- `AdHocEmailSendJob.cs` — 6 segment-filter queries
- `AuthorizedContactService.cs` — IsVetted projection in principal list
- `ReportService.cs` — vetted members CountAsync filter (reviewer find)

## In-memory service fixes (9 files)

All `(int)` and `(VettingStatus)` casts introduced as temporary bridges
in Phase 2 and Phase 3b-1 now removed — entity and DTO/request types
match, so direct assignments work:

- `UserManagementService.cs`, `MemberDetailsService.cs`, `UserDto.cs`
  constructor, `UserDashboardProfileService.cs`, `VettingService.cs`,
  `VettingSeeder.cs`, `TestHelperService.cs`
- `ProxyRsvpService.cs` — 2 magic-number `!= 3` comparisons → enum
- `CheckInService.cs` — walk-in attendee initializer (newly found
  during the Phase 3b-2 grep sweep)

## VettingHoldService.cs — full rewrite of bridge code

The service maintained 3 private integer constants
(`VettingStatus_FinalReview`, `_Approved`, `_OnHold`) and a
`GetStatusName(int)` helper as its own translation bridge. All
removed:

- The 3 `private const int VettingStatus_* = N;` lines deleted
- All comparisons (`user.VettingStatus != VettingStatus_Approved`,
  etc.) rewritten to use enum literals directly
- `GetStatusName(int)` signature changed to `GetStatusName(VettingStatus)`.
  The old int-based switch had drifted badly: it returned "NotStarted"
  for value 0 (actually UnderReview), "Submitted" for 1 (actually
  InterviewApproved), and had cases for values 7/8 that don't exist
  in the real enum. Rewrote using enum literals so the compiler
  enforces correctness going forward.
- DTO construction sites use enum literals directly; removed the
  cast in VettingHoldStatusResponse construction.

## Seed data (2 files, 26 sites)

- `UserSeeder.cs` — 24 VettingStatus initializer sites converted
  from int literals to enum values, including anonymous-type seed
  account records (lines 346-357). Final assignment in the foreach
  loop (line 412) now just copies the enum value without any cast.

- `AttendanceSeeder.cs` — **LATENT BUG FIX**. Two LINQ queries used
  `u.VettingStatus >= 3` as the "vetted users" filter. That `>= 3`
  range accidentally included Denied (4), OnHold (5), and Withdrawn
  (6) — exactly the users who should NOT be selected for social event
  RSVP seeding. Fixed to strict `== VettingStatus.Approved`. Dev/test
  seeder only (not production data), low severity but worth fixing
  while the file is being touched anyway.

## Test updates (20 files — done by test-executor agent)

The test-executor agent ran the full .NET test suite via the
run-test-suite skill and fixed every compile error caused by the entity
type change. The updates are mechanical int → enum conversions:

- DTO assertions: `.Should().Be(3)` → `.Should().Be(VettingStatus.Approved)`
- Entity initializers: `new ApplicationUser { VettingStatus = 3 }` →
  `= VettingStatus.Approved`
- Helper signatures: `CreateTestUser(int vettingStatus = 0, ...)` →
  `(VettingStatus vettingStatus = VettingStatus.UnderReview, ...)`
- The private `const int VettingStatus_*` aliases in
  VettingHoldServiceTests.cs converted to `const VettingStatus`

Files updated (full list in the test-executor agent's report):
- 11 unit test files (Auth, Dashboard, EmailTemplates,
  TicketAssignment × 4, Users × 1, Vetting, EmailTemplateServiceTrigger,
  MemberDetailsSecurityIntegration)
- 1 Core test file (VettingHoldServiceTests.cs)
- 8 integration test files (EmailTemplates, Participation, Vetting,
  VettingHold, TicketAssignment × 5)

### Test removal

`[InlineData(99, "Not Started")]` in
`MemberDetailsServiceBusinessLogicTests.GetMemberDetails_HandlesAllVettingStatusEnums`
was removed because the new CHECK constraint blocks value 99 at the DB
insert step — the test's "invalid value falls through to default" path
is now unreachable through persistence. The test now covers only the 7
valid enum values. A block comment explaining the removal was added.

### Known dead code (flagged, not fixed)

`MemberDetailsService.cs:122` has a `_ => "Not Started"` fallback in the
VettingStatusDisplay switch. With the new CHECK constraint it's
unreachable. Left in place for defense-in-depth; marked as a P3 item
to review in the future. Not a code fix target for Phase 3b-2.

## Verification

- `dotnet build`: 0 errors, 75 pre-existing warnings
- EF migration applied cleanly to dev database:
  * CK_Users_VettingStatus_Range constraint in place
  * VettingStatus distribution unchanged: {0: 5, 1: 1, 3: 12, 4: 2, 5: 2}
    (matches pre-migration distribution — no data loss/corruption)
- `restart-dev-containers` skill: 3/3 containers healthy
- `npm run generate` in packages/shared-types: no diff from previous
  generation (Phase 3b-1 already had the correct DTO shapes)
- `tsc --noEmit` in apps/web: 0 errors in Phase 3 files
- Browser verification on dev as applicant9@example.com (NervousNewbie,
  InterviewApproved): /dashboard alert renders correctly with
  calendar icon, "Great News!" title, working scheduling link
- Browser verification as admin@witchcityrope.com: /admin/members
  table renders all badges with historical labels and colors intact
- `run-test-suite` delegation to test-executor agent:
  * Aggregate: 1354 passed / 120 failed / 40 skipped (vs Phase 3b-1
    baseline 1342/127/40 — +12 pass, -7 fail)
  * **Phase 3b-2 regressions: 0**
  * All 120 remaining failures are pre-existing and tracked in
    docs/technical-debt.md (T-1 WAF pollution, T-2 EmailTemplate drift,
    Auth/Events/Volunteers/Metadata unrelated)

## Commits to main this phase

- 91d51770 Phase 3a — docblock + Withdrawn case
- cff87f33 Phase 3b-1 — DTO int → enum
- 7eaa7700 Phase 3b-1 code review S1 follow-up
- THIS     Phase 3b-2 — entity int → enum (ATOMIC)

## Scope discipline

45 files in this commit. Testing agent work is done, so I had permission
to update test files. No manual test runner invocations — all test
execution went through the run-test-suite skill via test-executor agent.

+++ apps/api/Features/EmailTemplates/Jobs/AdHocEmailSendJob.cs
@@ -286,11 +286,11 @@ public class AdHocEmailSendJob
         var filtered = segment switch
         {
             UserSegment.AllVettedMembers =>
-                query.Where(u => u.VettingStatus == (int)VettingStatus.Approved && u.IsActive),
+                query.Where(u => u.VettingStatus == VettingStatus.Approved && u.IsActive),
 
             UserSegment.AllPreVettedMembers =>
-                query.Where(u => u.VettingStatus != (int)VettingStatus.Denied
-                    && u.VettingStatus != (int)VettingStatus.OnHold
+                query.Where(u => u.VettingStatus != VettingStatus.Denied
+                    && u.VettingStatus != VettingStatus.OnHold
                     && u.IsActive),
 
             UserSegment.AllTeachers =>
@@ -309,10 +309,10 @@ public class AdHocEmailSendJob
                 query.Where(u => !u.EmailConfirmed && u.IsActive),
 
             UserSegment.VettingPending =>
-                query.Where(u => u.VettingStatus == (int)VettingStatus.UnderReview && u.IsActive),
+                query.Where(u => u.VettingStatus == VettingStatus.UnderReview && u.IsActive),
 
             UserSegment.NewImportedUsers =>
-                query.Where(u => u.VettingStatus == (int)VettingStatus.Approved
+                query.Where(u => u.VettingStatus == VettingStatus.Approved
                     && !u.EmailConfirmed
                     && u.IsActive),
